Lagt til bekreftelsesmelding ved utlogging og sletting av bruker.

+++ coaching/www/js/profile.js
@@ -4,8 +4,19 @@ readLogin();
 $("#user-email").val(login.email);
 $("#user-account").val(login.accountId);
 
+$("#deleteBtn").click(function(){
+    x = confirm('Er du sikker på at du vil slettte brukeren?');
+    if(x == true){
+        localStorage.clear();
+        window.location = '../index.html';
+    }
+});
+
 $("#logoutBtn").click(function(){
-    console.log('Logout');
-    localStorage.clear();
-    window.location = '../index.html';
+    x = confirm('Er du sikker på at du vil logge ut?');
+    if(x == true){
+        console.log('Logout');
+        localStorage.clear();
+        window.location = '../index.html';
+    }
 });
